Reorganization
Moved Rock-Paper-Scissors stuff off to a separate listener.

Moved coinflip command to an actual slash command, and upserted to Global level.
Moved RPS command to Global level as well.

Moved certain important information to an external config text file.
