+++ ChromeBotDriver.java
@@ -5,44 +5,55 @@
 import karuta.ChromeBotKarutaListener;
 import net.dv8tion.jda.api.JDA;
 import net.dv8tion.jda.api.JDABuilder;
+import net.dv8tion.jda.api.interactions.commands.Command;
 import net.dv8tion.jda.api.interactions.commands.OptionType;
 
 import javax.security.auth.login.LoginException;
+import java.io.File;
+import java.io.FileInputStream;
+import java.io.FileNotFoundException;
+import java.util.Scanner;
 
 public class ChromeBotDriver
 {
-    public static void main(String[] args) throws LoginException, InterruptedException
+    public static void main(String[] args) throws LoginException, InterruptedException, FileNotFoundException
     {
-        //  TEST SERVER (ID)
-        //  295301099512922114
+        final String TEST_SERVER_ID;
+        final String FRANKY_BOT_TOKEN, CHROME_BOT_TOKEN;
 
-        //  FRANKY BOT TOKEN:
-        //  ODg3ODAwNDcxNDk3MTEzNjkx.GM6zmn.RSv1awKuAwjGkMXRpunL2vzsI6WNwLcEmHvvcU
+        //  NOTE: Read important information like tokens from config file.
+        Scanner configReader = new Scanner(new File("config.txt"));
 
-        //  CHROME BOT TOKEN:
-        //  OTYwMjk1NDQxMTQ3MjYwOTk5.YkoW0g.8WryDhDx67_sHwWXWsr7a74zWnc
+        TEST_SERVER_ID = configReader.next().substring(15);
 
-        final String TEST_SERVER_ID = "295301099512922114";
-
-        final String FRANKY_BOT_TOKEN = "ODg3ODAwNDcxNDk3MTEzNjkx.GM6zmn.RSv1awKuAwjGkMXRpunL2vzsI6WNwLcEmHvvcU";
-        final String CHROME_BOT_TOKEN = "OTYwMjk1NDQxMTQ3MjYwOTk5.YkoW0g.8WryDhDx67_sHwWXWsr7a74zWnc";
+        FRANKY_BOT_TOKEN = configReader.next().substring(17);
+        CHROME_BOT_TOKEN = configReader.next().substring(17);
 
         //  BOT being BUILT
         JDA jda = JDABuilder.createDefault(FRANKY_BOT_TOKEN)
                 //.enableIntents(GatewayIntent.GUILD_MEMBERS)
                 //.setMemberCachePolicy(MemberCachePolicy.ALL)
                 .addEventListeners(new ChromeBotListener())
+                .addEventListeners(new ChromeBotRPSListener())
                 .addEventListeners(new ChromeBotKarutaListener())
                 .build();
 
         //  WAIT FOR JDA OBJECT TO ACTUALLY BE READY BEFORE DOING ANYTHING.
         jda.awaitReady();
 
-        //  Guild testServerGuild = jda.getGuildById(TEST_SERVER_ID);
-        jda.getGuildById(TEST_SERVER_ID).upsertCommand("rockpaperscissors", "Challenge someone to a game of Rock-Paper-Scissors!")
+        /*
+        jda.upsertCommand("rockpaperscissors", "Challenge someone to Rock-Paper-Scissors!")
                 .addOption(OptionType.USER, "opponent", "Username of the opponent.", true)
-                    .queue();
+                .queue();
+        jda.upsertCommand("coinflip", "Flip a coin!")
+                .queue();
 
-
+         */
+        
+        //  DEBUG: Display all commands the bot has in the TEST SERVER.
+        jda.getGuildById(TEST_SERVER_ID).retrieveCommands().queue(commands -> {
+            for(Command command : commands)
+                System.out.println(command.getName() + " " + command.getApplicationId() + " " + command.getId());
+        });
     }
 }
